refactor: split public dev-loop routing seams
